requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
psycopg2==2.9.4
PyMySQL==1.0.2
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.5
whitenoise==6.2.0
django-cors-headers==3.9.0